#42 feat : 로그인, 회원가입시 memberId 로컬 저장 로직 구현. MainViewModel 에서 꺼내서 채팅에 이용

+++ app/src/main/java/com/aoztg/greengrim/config/BearerInterceptor.kt
@@ -54,7 +54,7 @@ class BearerInterceptor : Interceptor {
                             sharedPreferences.edit()
                                 .putString(X_ACCESS_TOKEN, body.accessToken)
                                 .putString(X_REFRESH_TOKEN, body.refreshToken)
-                                .putInt(MEMBER_ID, body.memberId)
+                                .putLong(MEMBER_ID, body.memberId)
                                 .apply()
 
                             isRefreshed = true
