+++ app/src/main/java/com/aoztg/greengrim/data/model/SignupResponse.kt
@@ -3,5 +3,5 @@ package com.aoztg.greengrim.data.model
 data class SignupResponse(
     val accessToken: String,
     val refreshToken: String,
-    val memberId: Int
+    val memberId: Long
 )
